fix: rename Avartar to Avatar1
